requirements.txt
fastapi
uvicorn
influxdb-client
python-dotenv
streamlit
requests
pydantic